LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/registration-basic.https-expected.txt


PASS Registering normal scope 
FAIL Registering scope with fragment assert_equals: A fragment should be removed from scope expected "https://localhost:9443/service-workers/service-worker/resources/registration/scope-with-fragment" but got "https://localhost:9443/service-workers/service-worker/resources/registration/normal"
PASS Registering same scope as the script directory 
